+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Changed
+- Rename `cctx` ('canvas context') to `vctx` ('visual context')
+
 ## [0.7.0] - 2020-12-17
 ### Added
 - Importing with CommonJS syntax.